+++ src/ebonite/ext/aiohttp/__init__.py
@@ -1,0 +1,3 @@
+from .server import AIOHTTPServer
+
+__all__ = ['AIOHTTPServer']